Fix logits overflowing, cherry picked from https://github.com/marian-nmt/marian-dev/commit/1a7435827716ff125ce8dcee462bb392110f8a0b

+++ src/translator/translator.h
@@ -98,7 +98,7 @@ public:
           graph->getBackend()->setGemmType(options_->get<std::string>("gemm-type"));
           graph->getBackend()->setQuantizeRange(options_->get<float>("quantize-range"));
         }
-        graph->reserveWorkspaceMB(options_->get<size_t>("workspace"));
+        graph->reserveWorkspaceMB(options_->get<int>("workspace"));
         graphs_[id] = graph;
 
 #if MMAP
@@ -300,7 +300,7 @@ public:
         graph->getBackend()->setGemmType(options_->get<std::string>("gemm-type"));
         graph->getBackend()->setQuantizeRange(options_->get<float>("quantize-range"));
       }
-      graph->reserveWorkspaceMB(options_->get<size_t>("workspace"));
+      graph->reserveWorkspaceMB(options_->get<int>("workspace"));
       graphs_.push_back(graph);
 
       auto scorers = createScorers(options_);
